backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
transformers>=4.46.0
torch>=2.6.0
pydantic>=2.10.0
python-multipart>=0.0.12
pytest  # Add this for testing
requests # Add this (required by TestClient)
httpx
python-dotenv>=1.0.0
